+++ src/UI/Blocks/blockcolor.h
@@ -18,6 +18,7 @@ private slots:
     void on_primariesBox_currentIndexChanged(int index);
     void on_trcBox_currentIndexChanged(int index);
     void on_rangeBox_currentIndexChanged(int index);
+    void presetTriggered();
     void on_actionsRGB_triggered();
     void on_actionBT709_triggered();
     void on_actionUHD10_triggered();