+++ src/UI/Blocks/blockloops.cpp
@@ -3,6 +3,9 @@
 BlockLoops::BlockLoops(MediaInfo *mediaInfo, QWidget *parent) :
     BlockContentWidget(mediaInfo,parent)
 {
+#ifdef QT_DEBUG
+    qDebug() << "Create Loops block";
+#endif
     setupUi(this);
 }