+++ src/UI/inputwidget.cpp
@@ -7,10 +7,14 @@
 InputWidget::InputWidget(FFmpeg *ff, int id, QWidget *parent) :
     QWidget(parent)
 {
+#ifdef QT_DEBUG
+    qDebug() << "Create Input Widget";
+#endif
     setupUi(this);
 
     _mediaInfo = new MediaInfo( ff, this);
     _mediaInfo->setId(id);
+    connect ( _mediaInfo, SIGNAL(changed()), this, SLOT (updateOptions()));
 
     // CREATE MENUS
     blocksMenu = new QMenu(this);
@@ -124,7 +128,6 @@ BlockBaseWidget *InputWidget::addBlock(BlockContentWidget *content, QAction *act
 void InputWidget::addNewParam(QString name, QString value)
 {
     //add a param and a value
-    qDebug() << "New Custom param: " + name + " " + value;
     BlockCustom *block = new BlockCustom( _mediaInfo, name, value );
     BlockBaseWidget *bw = new BlockBaseWidget( "FFmpeg parameter", block, blocksWidget );
     blocksLayout->addWidget( bw );
@@ -170,6 +173,7 @@ void InputWidget::updateOptions()
     }
 
 
+    // After Effects options
     if (_mediaInfo->isAep())
     {
         actionAfter_Effects_composition->setVisible(true);
@@ -181,8 +185,23 @@ void InputWidget::updateOptions()
         actionAfter_Effects_composition->setVisible(false);
         blockAEComp->hide();
         actionAfter_Effects_threads->setVisible(false);
+        blockAEThreads->hide();
     }
-    blockAEThreads->hide();
+
+    //Color options
+    if ( _mediaInfo->hasVideo() )
+    {
+        VideoInfo *stream = _mediaInfo->videoStreams()[0];
+        if ( stream->colorTRC()->name() != "" || stream->colorPrimaries()->name() != "" || stream->colorRange()->name() != "" || stream->colorSpace()->name() != "")
+        {
+            blockColor->show();
+        }
+        else
+        {
+            blockColor->hide();
+        }
+    }
+
 }
 
 void InputWidget::updateCustomParams()